Submission of Project
Hi,

I m sorry for late submission but i had already talked with Shreya on
this, and took the time till Tuesday since i had my other interviews
scheduled until the last Wednesday.

For the project, i have tried to implement as much functionality as i
could within this time span; hence, i have not focused much on the UI
part. There were a couple of issues which needed more time to get
rectified, and hence i think i could have done much better than this if
i had
some more time.

Anyways, I did not have any experience or knowledge in angular JS pior
to this and the training and this prject was a great starting point.

Thank you for the training and this project. I learnt a lot from this.